// packages/api/lib/__generated__/graphql.d.ts
export type Maybe<T> = T | null;
export type RequireFields<T, K extends keyof T> = { [X in Exclude<keyof T, K>]?: T[X] } & { [P in K]-?: NonNullable<T[P]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string,
  String: string,
  Boolean: boolean,
  Int: number,
  Float: number,
  /** The `Upload` scalar type represents a file upload. */
  Upload: any,
};


export enum CacheControlScope {
  Public = 'PUBLIC',
  Private = 'PRIVATE'
}

export type Mutation = {
   __typename?: 'Mutation',
  _empty: Maybe<Scalars['String']>,
  addTrack: Array<Track>,
};


export type MutationAddTrackArgs = {
  input: TrackInput
};

export enum PlayerState {
  Playing = 'Playing',
  Stopped = 'Stopped',
  Paused = 'Paused',
  UnknownState = 'UnknownState'
}

export type Query = {
   __typename?: 'Query',
  _empty: Maybe<Scalars['String']>,
  playerState: PlayerState,
  currentTrack: Maybe<Track>,
  currentQueue: Array<Track>,
};

export type Subscription = {
   __typename?: 'Subscription',
  trackAdded: Track,
};

export type Track = {
   __typename?: 'Track',
  album: Scalars['String'],
  artist: Scalars['String'],
  cover: Maybe<Scalars['String']>,
  id: Scalars['ID'],
  title: Scalars['String'],
};

export type TrackInput = {
  title: Scalars['String'],
  artist: Scalars['String'],
  album: Scalars['String'],
};




export type ResolverTypeWrapper<T> = Promise<T> | T;

export type ResolverFn<TResult, TParent, TContext, TArgs> = (
  parent: TParent,
  args: TArgs,
  context: TContext,
  info: GraphQLResolveInfo
) => Promise<TResult> | TResult;


export type StitchingResolver<TResult, TParent, TContext, TArgs> = {
  fragment: string;
  resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};

export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> =
  | ResolverFn<TResult, TParent, TContext, TArgs>
  | StitchingResolver<TResult, TParent, TContext, TArgs>;
// ...
export type SubscriptionObject<TResult, TKey extends string, TParent, TContext, TArgs> =
  | SubscriptionSubscriberObject<TResult, TKey, TParent, TContext, TArgs>
  | SubscriptionResolverObject<TResult, TParent, TContext, TArgs>;

export type SubscriptionResolver<TResult, TKey extends string, TParent = {}, TContext = {}, TArgs = {}> =
  | ((...args: any[]) => SubscriptionObject<TResult, TKey, TParent, TContext, TArgs>)
  | SubscriptionObject<TResult, TKey, TParent, TContext, TArgs>;
// ...
export type ResolversTypes = {
  Query: ResolverTypeWrapper<{}>,
  String: ResolverTypeWrapper<Scalars['String']>,
  PlayerState: PlayerState,
  Track: ResolverTypeWrapper<Track>,
  ID: ResolverTypeWrapper<Scalars['ID']>,
  Mutation: ResolverTypeWrapper<{}>,
  TrackInput: TrackInput,
  Subscription: ResolverTypeWrapper<{}>,
  Boolean: ResolverTypeWrapper<Scalars['Boolean']>,
  CacheControlScope: CacheControlScope,
  Upload: ResolverTypeWrapper<Scalars['Upload']>,
  Int: ResolverTypeWrapper<Scalars['Int']>,
};

/** Mapping between all available schema types and the resolvers parents */
export type ResolversParentTypes = {
  Query: {},
  String: Scalars['String'],
  PlayerState: PlayerState,
  Track: Track,
  ID: Scalars['ID'],
  Mutation: {},
  TrackInput: TrackInput,
  Subscription: {},
  Boolean: Scalars['Boolean'],
  CacheControlScope: CacheControlScope,
  Upload: Scalars['Upload'],
  Int: Scalars['Int'],
};

export type CacheControlDirectiveResolver<Result, Parent, ContextType = any, Args = {   maxAge: Maybe<Maybe<Scalars['Int']>>,
  scope: Maybe<Maybe<CacheControlScope>> }> = DirectiveResolverFn<Result, Parent, ContextType, Args>;

export type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
  _empty: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>,
  addTrack: Resolver<Array<ResolversTypes['Track']>, ParentType, ContextType, RequireFields<MutationAddTrackArgs, 'input'>>,
};

export type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
  _empty: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>,
  playerState: Resolver<ResolversTypes['PlayerState'], ParentType, ContextType>,
  currentTrack: Resolver<Maybe<ResolversTypes['Track']>, ParentType, ContextType>,
  currentQueue: Resolver<Array<ResolversTypes['Track']>, ParentType, ContextType>,
};

export type SubscriptionResolvers<ContextType = any, ParentType extends ResolversParentTypes['Subscription'] = ResolversParentTypes['Subscription']> = {
  trackAdded: SubscriptionResolver<ResolversTypes['Track'], "trackAdded", ParentType, ContextType>,
};